moved menu nav to hook
