+++ src/components/common/Extensions/Navigation/Navigation.tsx
@@ -1,13 +1,11 @@
 import React, { FC, useState } from 'react';
-import { useColonyContext, useMobile } from '~hooks';
+import { useMobile, useSideNavigation } from '~hooks';
 import NavItem from './partials/NavItem';
 import Select from '~shared/Extensions/Select';
 
 const displayName = 'common.Extensions.Navigation';
 
 const Navigation: FC = () => {
-  const { colony } = useColonyContext();
-  const { name } = colony || {};
   const isMobile = useMobile();
 
   const [selectedElement, setSelectedElement] = useState(0);
@@ -16,50 +14,7 @@ const Navigation: FC = () => {
     setSelectedElement(selectedOption);
   };
 
-  const navigationItems = [
-    {
-      id: 0,
-      linkTo: `/colony/${name}`,
-      label: 'Colony Details',
-      value: 'colony-details',
-    },
-    {
-      id: 1,
-      linkTo: `/colony/${name}/reputation`,
-      label: 'Reputation',
-      value: 'reputation',
-    },
-    {
-      id: 2,
-      linkTo: `/colony/${name}/permissions`,
-      label: 'Permissions',
-      value: 'permissions',
-    },
-    {
-      id: 3,
-      linkTo: `/colony/${name}/extensions`,
-      label: 'Extensions',
-      value: 'extensions',
-    },
-    {
-      id: 4,
-      linkTo: `/colony/${name}/integrations`,
-      label: 'Integrations',
-      value: 'integrations',
-    },
-    {
-      id: 5,
-      linkTo: `/colony/${name}/incorporation`,
-      label: 'Incorporation',
-      value: 'incorporation',
-    },
-    {
-      id: 6,
-      linkTo: `/colony/${name}/advanced`,
-      label: 'Advanced',
-      value: 'advanced',
-    },
-  ];
+  const navigationItems = useSideNavigation();
 
   return (
     <nav role="navigation" className="flex flex-col w-full">
