+++ src/hooks/index.ts
@@ -52,6 +52,7 @@ export { default as useTokenTotalBalance } from './useTokenTotalBalance';
 export { default as useColonyContractVersion } from './useColonyContractVersion';
 export { default as useNetworkInverseFee } from './useNetworkInverseFee';
 export { default as useDetectClickOutside } from './useDetectClickOutside';
+export { default as useSideNavigation } from './useSideNavigation';
 
 /* Used in cases where we need to memoize the transformed output of any data.
  * Transform function has to be pure, obviously
